Randomly permute some some variables and run the OSA model on the permuted data

Randomly permute some some variables and run the OSA model on the permuted data

Usage

run_OSA_permuted(
  permutation_seeds = 1:100,
  permuting_variables = "Gender",
  patient_data,
  ...
)

Arguments

permutation_seeds

random number generator seeds to use to permute the data

permuting_variables

which variables to permute

patient_data

patient biomarker data

...

Arguments passed on to run_OSA

prob_score

array probability of each score for all subjects across all biomarkers

dim = number of subjects x number of biomarkers x number of scores

score_vals

a matrix specifying the scores for each biomarker

dim: number of biomarkers x number of scores

SuStaInLabels

the names of the biomarkers as a list of strings

N_startpoints

number of startpoints to use in maximum likelihood step of SuStaIn, typically 25

N_S_max

maximum number of subtypes, should be 1 or more

N_iterations_MCMC

number of MCMC iterations, typically 1e5 or 1e6 but can be lower for debugging

output_folder

where to save pickle files, etc.

dataset_name

for naming pickle files

use_parallel_startpoints

boolean for whether or not to parallelize the maximum likelihood loop

seed

random number seed for python code

plot

logical() indicating whether to construct PVD plots via python subroutines

N_CV_folds

number of cross-validation folds to create

CV_fold_nums

which CV folds to run (for parallel processing)

verbose

logical() indicating whether to print debugging information

keep_data

logical() indicating whether to include the ata in the return object

fig_size

python figure size, in inches (width, height)

prob_correct

the probability of correctly classifying the underlying biomarker level: p(obs = true)

biomarker_levels

a list containing the levels for each biomarker

Value

NULL, invisibly (.pickle and .rds files are generated to save the output)